cosmetics

+++ src/main/java/org/primefaces/component/treetable/TreeTable.java
@@ -268,7 +268,7 @@ public class TreeTable extends TreeTableBase {
             else if (eventName.equals("page")) {
                 int rows = getRowsToRender();
                 int first = Integer.parseInt(params.get(clientId + "_first"));
-                int page = rows > 0 ? (int) (first / rows) : 0;
+                int page = rows > 0 ? (first / rows) : 0;
 
                 wrapperEvent = new PageEvent(this, behaviorEvent.getBehavior(), page);
                 wrapperEvent.setPhaseId(behaviorEvent.getPhaseId());
